visits added

+++ src/main/java/com/andrew/service/DeliveryService.java
@@ -5,12 +5,14 @@ import com.andrew.dto.delivery.DeliveryResponseDTO;
 import com.andrew.exceptions.NotFoundException;
 import com.andrew.mapper.dto.DeliveryMapper;
 import com.andrew.model.Delivery;
+import com.andrew.model.VisitRequest;
+import com.andrew.model.enums.VisitRequestStatus;
 import com.andrew.repository.DeliveryRepository;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import jakarta.security.enterprise.SecurityContext;
 import jakarta.transaction.Transactional;
+import jakarta.ws.rs.ForbiddenException;
 
 @ApplicationScoped
 public class DeliveryService {
@@ -21,34 +23,48 @@ public class DeliveryService {
     DeliveryMapper deliveryMapper;
 
     @Inject
-    private SecurityContext securityContext;
+    VisitRequestService visitRequestService;
+
+    @Transactional
+    public Delivery getById(Long id) {
+        return deliveryRepository.findById(id)
+            .orElseThrow(() -> new NotFoundException("Delivery", id));
+    }
 
     @Transactional
     public DeliveryResponseDTO createDelivery(DeliveryCreateDTO dto) {
+        VisitRequest visitRequest = visitRequestService.getById(dto.visitRequestId());
+        if (!visitRequest.getStatus().equals(VisitRequestStatus.DRAFT))
+            throw new ForbiddenException();
+        
         return createDelivery(deliveryMapper.toEntity(dto));
     }
 
     @Transactional
-    public DeliveryResponseDTO createDelivery(Delivery delivery) {
+    private DeliveryResponseDTO createDelivery(Delivery delivery) {
         deliveryRepository.save(delivery);
         return deliveryMapper.toResponse(delivery);
     }
 
     @Transactional
     public DeliveryResponseDTO updateDelivery(Long id, DeliveryCreateDTO dto) {
-        deliveryRepository.findById(id)
-            .orElseThrow(() -> new NotFoundException("Delivery", id));
+        Delivery existing = getById(id);
+        if (!existing.getVisitRequest().getStatus().equals(VisitRequestStatus.DRAFT))
+            throw new ForbiddenException();
 
         Delivery toUpdate = deliveryMapper.toEntity(dto);
         toUpdate.setId(id);
+        if (!toUpdate.getVisitRequest().getId().equals(existing.getVisitRequest().getId()))
+            throw new ForbiddenException();
 
         return deliveryMapper.toResponse(deliveryRepository.update(toUpdate));
     }
 
     @Transactional
     public void deleteDelivery(Long id) {
-        Delivery delivery = deliveryRepository.findById(id)
-            .orElseThrow(() -> new NotFoundException("Delivery", id));
+        Delivery delivery = getById(id);
+        if (!delivery.getVisitRequest().getStatus().equals(VisitRequestStatus.DRAFT))
+            throw new ForbiddenException();
 
         deliveryRepository.delete(delivery);
     }
